Fix import error of `sila_cetoni.motioncontrol`
due to import error of `Server`

Move the import of `Server` to `create_server` to allow successful
import of `CetoniAxisSystemDevice` from other modules even if the import
of `Server` will fail later. This is to allow other modules
(`sila_cetoni.application.application_system`) to import this module for
the types it defines.